Fix for Issue #88: Moved test functions for tardis.util from test_config_reader to test_util
